Kitchen: started in on the kitchen
? This included renaming the various documentary images so that they're 
easier to work from as references.

? As I make successive rooms I'm getting really, really sick of how hard 
it is to move between Pixen and Phaser in terms of the correspondences, 
it's just awwwwwful. AWFUL. Surely there is a better way? Maybe "Export 
Layers" and just allow everything to be full size, knowing TexturePacker 
will trim, and then can position everything at center and only worry 
about collision boxes for setting depth? Something like that.