Added Atrium scene with table and chairs.
? Crucial realization as I struggled to work out how to be able to 
trigger a message when Marina hits the tape outline as per the first 
game: Marina doesn't need to obey the tape!! She is the Artist, it's her 
domain, she can just step over it at will. A tiny little difference in 
implementation, but to me seems emblematic of a crucial way in which 
this game is different - Marina is more empowered here, but less in 
other ways?