Basics of a dialog system implemented.
? Lots of technical struggles on this one, hugely caused by not knowing 
that a Group and a Container behave very differently in Phaser. Not the 
most interesting experience to struggle with for quite so long.

? Having some difficulty recreating the precise look of the original 
game - the font seems to render heavier when I'm using it in this game 
for some reason I don't quite understand. It bothers me a little for the 
game not to be identical, but I may have to lump it.

? As I type I'm realising it'll be fun and funny to have a message 
mimicking the original opening message when you arrive at MoMA.