Significant progress on generating new people
? Working with Colbydude's palette swapping code from 
https://github.com/Colbydude/phaser-3-palette-swapping-example which is 
reasonably straightforward. No idea as yet about performance, but it 
seems to be non-terrible, so I'm assuming we'll be okay.

? Managed to repurpose a bunch of the values/code from the original game 
which has my life somewhat easier, though having some odd trouble with 
the "long hair" part of people's head. Presumably there's a sensible 
solution, but I don't know it yet. Anyway, the overall direction is a 
good one, with random-looking people being generated not too poorly!