Have the basics of the gaze cycle working rather suddenly.
? There's a ton of timing niceties (and perhaps randomization for 
"organic" time) that need to be put in so it doesn't feel too much like 
a machine?

? But I'm quite surprised by how that suddenly jumped together, and I 
does feel like it has something when you're looking at someone.

? I'm finding the dialogs that tell you about looking up and down useful 
but obtrusive which is a pain. I wonder if there's another way of 
designing this that doesn't end up occluding the whole screen so you 
could spend a bit more time being able to just look at Marina with her 
head down?

? Plausibly I could just move the location of the box for these sorts of 
message? Just move it to the top of bottom?

? Anyway, it's kind of functionally complete now, sans crying and 
blinking, and an end of day concept.