Extremely basic crying
? I'm surprised that it's as effective (affective) as it is. I think 
it'll likely work this way. There's a debate to be had around all the 
more organic/realism elements. Should tears catch on the chin before 
falling? Should they fall at all? Should they move at different speeds? 
How totally asymmetric?

? Probably the "answer" here is to think in terms of symbols instead of 
realisms. It's about making it clear the person is crying, not 
completely studying their tears? Also it's not an always on thing, so 
it's not something I'm relying on for visual interest (though it'd 
provide it). That said, the thought of it being the only thing a player 
might be able to watch (other than blinking (to come) makes me wonder if 
a little variation might go a long way?

? And as always the tension between literal resolution (low) and the 
resolution of reality (the more organic the behaviour the more uncanny 
the overall effect?)