Added dialog for bumping into queue and guards
? Definitely feels like a more dynamic space with all this. On the other 
hand, the writing ends up being pretty insipid. I wonder what kind of 
balancing act is needed here. The dialog is almost symbolic - it's about 
your function in the system of the work - but if it's too symbolic maybe 
it ends up as trite rather than drawing you official into your role? 
It's the ultimate escalation of your significance prior to your task.